/go/performance/utils/sysbench_runner: wip, still refactoring
